fix a rare race condition with activerecord cache
the issue
---------
in some cases, there is a race condition between checkine if a key
already exists in the database and inserting that key.
When this happens, activerecord the database adapter throw an exception
because it tries to insert a key that already exists

the fix
-------
it rescue from the exception and throw a simple warning.
it also introduce a "silence_warnings" configuration for Copycat.